backend/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
fairlearn
shap
python-dotenv
google-genai
python-multipart
folktables